requirements.txt
vrchatapi
aiohttp
python-dotenv
